+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2020-01-23 1.10.60
+  - Find Missing Thumbnails on startup
+  - Improved speed on RGBW brightness change
+  - LED offset of 0 is correctly changed/displayed
 # 2020-01-02 1.10.57-59
   - Fixed up regenerate_thumbnails()
   - Table is aware of number of thumbnails in generation queue